requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
PyJWT==2.8.0
bcrypt==4.1.2
marshmallow==3.20.2
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
email-validator==2.1.0
phonenumbers==8.13.26